refactor: json types
