+++ Typeform/Json/Serialization/JsonSnakeCaseNamingPolicy.cs
@@ -1,4 +1,4 @@
-namespace System.Text.Json;
+namespace System.Text.Json.Serialization;
 
 /// <summary>
 /// https://github.com/J0rgeSerran0/JsonNamingPolicy/blob/3400da9e73317d463178012ca5f955b412bad5b4/JsonSnakeCaseNamingPolicy.cs
