+++ Typeform/Json/Serialization/TypeformVariableJsonConverter.cs
@@ -1,7 +1,7 @@
 
 using System.Text.Json.Serialization;
 
-namespace Typeform.Serialization;
+namespace Typeform.Json.Serialization;
 
 public class TypeformVariableJsonConverter : JsonConverter<TypeformVariable>
 {
